Moved Suggesting to Recipe Book Class and Testing - Bug in Testing
